Assigment Html & CSS
I am attaching the work done in this way, since I get an error when trying to push the files. I will try to solve the problem, but first I wanted to send you the work done.

The html document "index.html" is the part of the structure and the css document "style.css" is the style that I have made.
I have tried to respect all the measures.

I hope it can be seen well, if there is any problem, do not hesitate to tell me, since I think it is quite good.